+++ python/katas/py7kyu/find_the_next_perfect_square.py
@@ -3,7 +3,7 @@
 import math
 
 
-def find_next_square(sq: int) -> int:
+def find_next_square(sq: int) -> float:
     r: float = math.sqrt(sq)
 
     return -1 if r % 1 else (r + 1) ** 2
